Wording clarification: "Verify" --> "Review"
When submitting a new ad, it may not be immediately clear what is meant by "verify". I believe the term "review" reflects more accurately and unambiguously what Step 2 entails.

This change also avoids having two meanings for the word "verify", which is also used in the context of confirming a previously unseen email address.