Add custom favicon to replace Jobberbase's default
Note: due to the simple geometry, the SVG favicon was written by hand,
to minimize the file size and maximize readability
(as much as the D attribute of SVG path elements allows, anyway.)

The shapes were drawn to align with a 16x16 grid.
The paths start at the top left corner of each letter,
and consist of horizontal and vertical line segments,
as well as and circular arcs for the J's curve.

A PNG version of the favicon is added as well, at 64x64px, for reference.