Fix display of job type markers
They were being created as block elements (`<div>` elements),
which are not permitted inside paragraphs (`<p>` elements).

This change makes them into `<span>` elements, which are valid there.
This avoids a browser-initiated reshuffling of the elements
which was breaking the layout of the job listings
(by making the tags appear on the line under the job details).