Reduce size and tweak colors of job-type bullets
Now that the colored bullet is inside the tag itself,
the diameter of the circle should be adjusted
to harmonize better with the text alongside it.

Additionally, the colors have been tweaked
to make them more distinguishable in hue (no longer there are two blues)
and less divergent in terms of luminance.